+++ src/components/feature.tsx
@@ -1,0 +1,78 @@
+import { Expand, Scissors, Target, Zap } from "lucide-react";
+import { Card, CardContent } from "./ui/card";
+
+  const features = [
+    {
+      icon: <Scissors className="h-8 w-8" />,
+      title: "AI Background Removal",
+      description:
+        "Remove backgrounds instantly with advanced AI technology. Perfect for product photos and portraits.",
+      color: "text-emerald-600",
+      bgColor: "bg-emerald-100",
+    },
+    {
+      icon: <Expand className="h-8 w-8" />,
+      title: "Smart Upscaling",
+      description:
+        "Enhance image quality and resolution without losing clarity using cutting-edge AI algorithms.",
+      color: "text-blue-600",
+      bgColor: "bg-blue-100",
+    },
+    {
+      icon: <Target className="h-8 w-8" />,
+      title: "Object-Focused Cropping",
+      description:
+        "Intelligently crop images around specific objects with AI-powered detection and framing.",
+      color: "text-purple-600",
+      bgColor: "bg-purple-100",
+    },
+    {
+      icon: <Zap className="h-8 w-8" />,
+      title: "Lightning Fast",
+      description:
+        "Process images in seconds, not minutes. Our optimized AI infrastructure delivers results instantly.",
+      color: "text-amber-600",
+      bgColor: "bg-amber-100",
+    },
+  ];
+
+const FeatureSection = () => {
+  return (
+    <section id="features" className="bg-white py-20 sm:py-32">
+        <div className="mx-auto max-w-7xl px-4 sm:px-6 lg:px-8">
+          <div className="mx-auto max-w-2xl text-center">
+            <h2 className="text-3xl font-bold tracking-tight text-slate-800 sm:text-4xl font-heading">
+              Powerful AI Tools at Your {" "}
+              <span className="bg-gradient-to-r from-blue-600 to-purple-600 bg-clip-text text-transparent font-heading">Fingertips</span>
+            </h2>
+            <p className="mt-4 text-lg text-slate-600 font-sans">
+              Everything you need to create stunning images with the power of
+              artificial intelligence
+            </p>
+          </div>
+
+          <div className="grid gap-8 sm:grid-cols-2 lg:grid-cols-4">
+            {features.map((feature,index) => (
+              <Card
+              key={index}
+              className="group relative overflow-hidden border-slate-200 bg-white/70 backdrop-blur-sm transition-all hover:shadow-lg"
+              >
+                <CardContent>
+                  <div 
+                  className={`${feature.bgColor} mb-4 inline-flex items-center justify-center rounded-lg p-3 ${feature.color}`}
+                  >
+                    {feature.icon}
+                  </div>
+                  <h3 className="text-lg mb-2 font-semibold text-slate-800 font-heading">{feature.title}</h3>
+                  <p className="text-sm text-slate-600 font-sans">{feature.description}</p>
+                </CardContent>
+                <div className="absolute inset-0 bg-gradient-to-r from-blue-500/0 to-blue-500/5 opacity-0 transition-opacity group-hover:opacity-100" />
+              </Card>
+            ))}
+          </div>
+        </div>
+      </section>
+  )
+}
+
+export default FeatureSection
